+++ R/clean_parameters.R
@@ -194,6 +194,11 @@ clean_parameters.wbm <- function(x, ...) {
 clean_parameters.wbgee <- clean_parameters.wbm
 
 
+#' @export
+clean_parameters.merModList <- function(x, ...) {
+  clean_parameters.default(x[[1]], ...)
+}
+
 
 #' @export
 clean_parameters.glmm <- function(x, ...) {
@@ -319,6 +324,30 @@ clean_parameters.afex_aov <- function(x, ...) {
   } else {
     clean_parameters(x$lm, ...)
   }
+}
+
+
+#' @export
+clean_parameters.mlm <- function(x, ...) {
+  pars <- find_parameters(x, effects = "all", component = "all", flatten = FALSE)
+
+  l <- lapply(names(pars), function(i) {
+    eff <- "fixed"
+    com <- "conditional"
+    data.frame(
+      Parameter = pars[[i]]$conditional,
+      Effects = eff,
+      Component = com,
+      Response = i,
+      Cleaned_Parameter = pars[[i]]$conditional,
+      stringsAsFactors = FALSE,
+      row.names = NULL
+    )
+  })
+
+  out <- .remove_backticks_from_parameter_names(do.call(rbind, l))
+  out <- .remove_empty_columns_from_pars(out)
+  .fix_random_effect_smooth(x, out)
 }