+++ R/get_weights.R
@@ -4,23 +4,25 @@
 #' @description Returns weighting variable of a model.
 #'
 #' @param x A fitted model.
+#' @param na_rm Logical, if \code{TRUE}, removes possible missing values.
 #' @param ... Currently not used.
 #'
-#' @return The weighting variable, or \code{NULL} if no weights were specified.
+#' @return The weighting variable, or \code{NULL} if no weights were specified (or if all weights were 1).
 #'
 #' @examples
 #' data(mtcars)
 #' mtcars$weight <- rnorm(nrow(mtcars), 1, .3)
 #' m <- lm(mpg ~ wt + cyl + vs, data = mtcars, weights = weight)
 #' get_weights(m)
+#' @importFrom stats na.omit
 #' @export
-get_weights <- function(x, ...) {
+get_weights <- function(x, na_rm = FALSE, ...) {
   UseMethod("get_weights")
 }
 
 
 #' @export
-get_weights.default <- function(x, ...) {
+get_weights.default <- function(x, na_rm = FALSE, ...) {
   w <- NULL
   tryCatch(
     {
@@ -71,15 +73,34 @@ get_weights.default <- function(x, ...) {
     )
   }
 
+  # if all weights are 1, set return value to NULL
+  if (!is.null(w) && all(w == 1L)) {
+    w <- NULL
+  }
+
+  if (!is.null(w) && anyNA(w) && isTRUE(na_rm)) {
+    w <- stats::na.omit(w)
+  }
+
   w
 }
 
 
 #' @export
-get_weights.brmsfit <- function(x, ...) {
+get_weights.brmsfit <- function(x, na_rm = FALSE, ...) {
   w <- find_weights(x)
 
   if (!is.null(w)) {
-    get_data(x)[[w]]
+    w <- get_data(x)[[w]]
   }
+
+  if (!is.null(w) && all(w == 1L)) {
+    w <- NULL
+  }
+
+  if (!is.null(w) && anyNA(w) && isTRUE(na_rm)) {
+    w <- stats::na.omit(w)
+  }
+
+  w
 }