+++ R/link_function.R
@@ -344,9 +344,24 @@ link_function.logitor <- link_function.logitmfx
 
 
 #' @export
+link_function.merModList <- function(x, ...) {
+  link_function.default(x[[1]], ...)
+}
+
+
+#' @export
 link_function.mipo <- function(x, ...) {
   models <- eval(x$call$object)
   link_function(models$analyses[[1]])
+}
+
+
+#' @export
+link_function.mira <- function(x, ...) {
+  if (!requireNamespace("mice", quietly = TRUE)) {
+    stop("Package `mice` required. Please install it.", call. = FALSE)
+  }
+  link_function(mice::pool(x), ...)
 }